Replace sin and cos with sinf and cosf for float

+++ src/NuMagSANSlib_gpuKernel.h
@@ -257,23 +257,23 @@ void AzimuthalAverage(ScatteringData SANSData){
        unsigned int N_r = *SANSData.N_r;
        unsigned int N_q = *SANSData.N_q;
        float dq = *SANSData.dq;
-       float qr1 = 0.0;
-       float qr2 = 0.0;
+       float qr1 = 0.0f;
+       float qr2 = 0.0f;
        bool b1 = false; 
        bool b2 = false; 
-       float s1 = 0.0;
-       float s2 = 0.0;
+       float s1 = 0.0f;
+       float s2 = 0.0f;
   
        if(i < N_r){
                for(int j=0; j<N_q-1; j++){
   
                    qr1 = SANSData.q_1D[j] * SANSData.r_1D[i];
                    b1 = (qr1 == 0.0f);
-                   s1 = (sin(qr1)/(qr1 + (float)b1) + (float)b1) * pow(SANSData.q_1D[j], 2);
+                   s1 = (sinf(qr1)/(qr1 + (float)b1) + (float)b1) * pow(SANSData.q_1D[j], 2);
                    
 	               qr2 = SANSData.q_1D[j+1] * SANSData.r_1D[i];
 	               b2 = (qr2 == 0.0f);
-        	       s2 = (sin(qr2)/(qr2 + (float)b2) + (float)b2) * pow(SANSData.q_1D[j+1], 2);
+        	       s2 = (sinf(qr2)/(qr2 + (float)b2) + (float)b2) * pow(SANSData.q_1D[j+1], 2);
 
              	   SANSData.c_Nuc_unpolarized[i] += SANSData.S_Nuc_1D_unpolarized[j]  * s1 \
              	   								  + SANSData.S_Nuc_1D_unpolarized[j+1] * s2;
@@ -320,11 +320,11 @@ void CorrelationFunction_2D(ScatteringData SANSData){
     int L_fourier = (*SANSData.N_q) * (*SANSData.N_theta);
     float v = 1.0/((float) L_fourier);
     int L_real = (*SANSData.N_r) * (*SANSData.N_alpha);
-    float c = 0.0;
+    float c = 0.0f;
 
     if(i < L_real){
         for(int k = 0; k<L_fourier; k++){
-            c = cos(SANSData.qy_2D[k] * SANSData.ry_2D[i] + SANSData.qz_2D[k] * SANSData.rz_2D[i]);
+            c = cosf(SANSData.qy_2D[k] * SANSData.ry_2D[i] + SANSData.qz_2D[k] * SANSData.rz_2D[i]);
             SANSData.Corr_Nuc_2D_unpolarized[i] += v * SANSData.S_Nuc_2D_unpolarized[k] * c;
             SANSData.Corr_Mag_2D_unpolarized[i] += v * SANSData.S_Mag_2D_unpolarized[k] * c;
             SANSData.Corr_NucMag_2D[i] += v * SANSData.S_NucMag_2D[k] * c;
@@ -459,8 +459,8 @@ void Atomistic_MagSANS_Kernel_dilute(MagnetizationData MagData,\
 				Psi = Y * SANSData.qy_2D[i] + Z * SANSData.qz_2D[i];
 
 				// cosine and sine values
-				cos_val = cos(Psi);
-				sin_val = sin(Psi);
+				cos_val = cosf(Psi);
+				sin_val = sinf(Psi);
 
 				// cosine and sine summations
             	mx_real += MagData.mx[l+N_cum] * cos_val;
@@ -481,8 +481,8 @@ void Atomistic_MagSANS_Kernel_dilute(MagnetizationData MagData,\
 			Mz_imag = MagData.RotMat[2] * mx_imag + MagData.RotMat[5] * my_imag + MagData.RotMat[8] * mz_imag;
 
 
-			cos_theta = cos(SANSData.theta_2D[i]);
-			sin_theta = sin(SANSData.theta_2D[i]);
+			cos_theta = cosf(SANSData.theta_2D[i]);
+			sin_theta = sinf(SANSData.theta_2D[i]);
 
 			// real-parts of the Halpern-Johnson vector
 			Qx_real = (-Mx_real);
@@ -618,8 +618,8 @@ void Atomistic_NucSANS_Kernel_dilute(NuclearData NucData,\
 				Psi = Y * SANSData.qy_2D[i] + Z * SANSData.qz_2D[i];
 
 				// cosine and sine values
-				cos_val = cos(Psi);
-				sin_val = sin(Psi);
+				cos_val = cosf(Psi);
+				sin_val = sinf(Psi);
 
 				// cosine- and sine-summations
 				Nuc_real += NucData.Nuc[l+N_cum] * cos_val;
@@ -790,8 +790,8 @@ void Atomistic_NuMagSANS_Kernel_dilute(NuclearData NucData,\
 			Mz_imag = MagData.RotMat[2] * mx_imag + MagData.RotMat[5] * my_imag + MagData.RotMat[8] * mz_imag;
 
 
-			cos_theta = cos(SANSData.theta_2D[i]);
-			sin_theta = sin(SANSData.theta_2D[i]);
+			cos_theta = cosf(SANSData.theta_2D[i]);
+			sin_theta = sinf(SANSData.theta_2D[i]);
 
 			// real-parts of the Halpern-Johnson vector
 			Qx_real = (-Mx_real);
@@ -973,8 +973,8 @@ void Atomistic_MagSANS_Kernel(MagnetizationData MagData,\
 				Psi = Y * SANSData.qy_2D[i] + Z * SANSData.qz_2D[i];
 
 				// cosine and sine values
-				cos_val = cos(Psi);
-				sin_val = sin(Psi);
+				cos_val = cosf(Psi);
+				sin_val = sinf(Psi);
 
 				// cosine and sine summations
             	mx_real += MagData.mx[l+N_cum] * cos_val;
@@ -996,8 +996,8 @@ void Atomistic_MagSANS_Kernel(MagnetizationData MagData,\
 
 		}
 
-		cos_theta = cos(SANSData.theta_2D[i]);
-		sin_theta = sin(SANSData.theta_2D[i]);
+		cos_theta = cosf(SANSData.theta_2D[i]);
+		sin_theta = sinf(SANSData.theta_2D[i]);
 
 		// real-parts of the Halpern-Johnson vector
 		Qx_real = (-Mx_real);
@@ -1139,8 +1139,8 @@ void Atomistic_NucSANS_Kernel(NuclearData NucData,\
 				Psi = Y * SANSData.qy_2D[i] + Z * SANSData.qz_2D[i];
 
 				// cosine and sine values
-				cos_val = cos(Psi);
-				sin_val = sin(Psi);
+				cos_val = cosf(Psi);
+				sin_val = sinf(Psi);
 
 				nuc_real += NucData.Nuc[l+N_cum] * cos_val;
 				nuc_imag -= NucData.Nuc[l+N_cum] * sin_val;
@@ -1281,8 +1281,8 @@ void Atomistic_NuMagSANS_Kernel(NuclearData NucData, \
 				Psi = Y * SANSData.qy_2D[i] + Z * SANSData.qz_2D[i];
 
 				// cosine and sine values
-				cos_val = cos(Psi);
-				sin_val = sin(Psi);
+				cos_val = cosf(Psi);
+				sin_val = sinf(Psi);
 
 				// cosine and sine summations
 				nuc_real += NucData.Nuc[l+N_cum] * cos_val;
@@ -1311,8 +1311,8 @@ void Atomistic_NuMagSANS_Kernel(NuclearData NucData, \
 
 		}
 
-		cos_theta = cos(SANSData.theta_2D[i]);
-		sin_theta = sin(SANSData.theta_2D[i]);
+		cos_theta = cosf(SANSData.theta_2D[i]);
+		sin_theta = sinf(SANSData.theta_2D[i]);
 
 		// real-parts of the Halpern-Johnson vector
 		Qx_real = (-Mx_real);
